tax-auto-demo-railway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
pytesseract==0.3.10
stripe==7.8.0
psycopg2-binary==2.9.9
alembic==1.13.1
email_validator